Extract the Joint Membership of a pair of Countries across IGOs.

Dyadic version of the data. The unit of observation is a dyad of countries. It provides a summary of the joint memberships of two IGOs over time.

igo_dyadic(country1, country2, year = 1816:2014, ioname = NULL)

Arguments

country1
A single state, used as a base of comparison. It could be any valid name or code of a state as specified on states2016.

country2
A state of vector of states to be compared with country1.

year
Year to be assessed, an integer or an array of year.

ioname
Optional. ioname or vector of ioname corresponding to the IGOs to be assessed. If NULL (the default), all IGOs would be extracted. See codes on igo_search.

Source

Codebook Version 3 IGO Data for full reference.

Value

A coded data frame representing the years and country dyad (rows) and the IGOs selected (columns). See Details

Details

This function tries to replicate the information contained in the original file distributed by The Correlates of War Project (dyadic_format3.dta). That file is not included in this package due to its size.

The result is a data frame containing the common years of the states selected via country1, country2, year by rows.

An additional column dyadid, computed as (1000*ccode1)+ccode2 is provided in order to identify relationships.

For each IGO selected via ioname (or all if the default option has been used) a column (using lowercase ioname as identifier) is provided with the following code system:

Category	Numerical Value

No Joint Membership	0

Joint Full Membership	1

Missing data	-9

State Not System Member	-1

If one state in an IGO is a full member but the other is an associate member or observer, that IGO is not coded as a joint membership.

Differences with the original dataset

There are some differences on the results provided by this function and the original dataset on some IGOs regarding the "Missing Data" (-9) and "State Not System Member" (-1). However it is not clear how to fully replicate those values.

See Codebook Version 3 IGO Data

References

Pevehouse, Jon CW, Timothy Nordstrom, Roseanne W McManus, and Anne Spencer Jamison. "Tracking Organizations in the World: The Correlates of War IGO Version 3.0 Datasets." Journal of Peace Research 57, no. 3 (May 2020): 492-503. doi: 10.1177/0022343319881175 .

See also

Examples

USA_CAN <- igo_dyadic("USA", "Spain")
nrow(USA_CAN)

#> [1] 199

ncol(USA_CAN)

#> [1] 544

# Using custom parameters
igo_dyadic(
  country1 = "France",
  country2 =
    c("Sweden", "Austria"),
  year = 1992:1995,
  ioname = "EU"
)

#>   dyadid ccode1 stateabb1 statenme1 state1 ccode2 stateabb2 statenme2  state2
#> 1 220380    220       FRN    France france    380       SWD    Sweden  sweden
#> 2 220380    220       FRN    France france    380       SWD    Sweden  sweden
#> 3 220380    220       FRN    France france    380       SWD    Sweden  sweden
#> 4 220380    220       FRN    France france    380       SWD    Sweden  sweden
#> 5 220305    220       FRN    France france    305       AUS   Austria austria
#> 6 220305    220       FRN    France france    305       AUS   Austria austria
#> 7 220305    220       FRN    France france    305       AUS   Austria austria
#> 8 220305    220       FRN    France france    305       AUS   Austria austria
#>   year eu
#> 1 1992 -1
#> 2 1993  0
#> 3 1994  0
#> 4 1995  1
#> 5 1992 -1
#> 6 1993  0
#> 7 1994  0
#> 8 1995  1