Extract Members of an IGO

Extract all the countries belonging to an IGO on a specific date.

igo_members(ioname, year = NULL, status = "Full Membership")

Arguments

ioname
Any valid ioname of an IGO as specified on igo_year_format3. It could be also a vector of IGOs.

year
Year to be assessed, an integer or an array of year. If NULL the latest year available of the IGO would be extracted.

status
Character or vector with the membership status to be extracted. See Details on state_year_format3.

Value

A dataframe.

See also

Examples

# Composition on two different dates
igo_members("EU", year = 1993)

#>    ioname ccode       state       statenme year value        category
#> 1      EU   211     belgium        Belgium 1993     1 Full Membership
#> 2      EU   390     denmark        Denmark 1993     1 Full Membership
#> 3      EU   220      france         France 1993     1 Full Membership
#> 4      EU   255     germany        Germany 1993     1 Full Membership
#> 5      EU   350      greece         Greece 1993     1 Full Membership
#> 6      EU   205     ireland        Ireland 1993     1 Full Membership
#> 7      EU   325       italy          Italy 1993     1 Full Membership
#> 8      EU   212  luxembourg     Luxembourg 1993     1 Full Membership
#> 9      EU   210 netherlands    Netherlands 1993     1 Full Membership
#> 10     EU   235    portugal       Portugal 1993     1 Full Membership
#> 11     EU   230       spain          Spain 1993     1 Full Membership
#> 12     EU   200          uk United Kingdom 1993     1 Full Membership
#>           orgname
#> 1  European Union
#> 2  European Union
#> 3  European Union
#> 4  European Union
#> 5  European Union
#> 6  European Union
#> 7  European Union
#> 8  European Union
#> 9  European Union
#> 10 European Union
#> 11 European Union
#> 12 European Union

igo_members("EU")

#>    ioname ccode       state       statenme year value        category
#> 1      EU   305     austria        Austria 2014     1 Full Membership
#> 2      EU   211     belgium        Belgium 2014     1 Full Membership
#> 3      EU   355    bulgaria       Bulgaria 2014     1 Full Membership
#> 4      EU   344     croatia        Croatia 2014     1 Full Membership
#> 5      EU   352      cyprus         Cyprus 2014     1 Full Membership
#> 6      EU   316    czechrep Czech Republic 2014     1 Full Membership
#> 7      EU   390     denmark        Denmark 2014     1 Full Membership
#> 8      EU   366     estonia        Estonia 2014     1 Full Membership
#> 9      EU   375     finland        Finland 2014     1 Full Membership
#> 10     EU   220      france         France 2014     1 Full Membership
#> 11     EU   255     germany        Germany 2014     1 Full Membership
#> 12     EU   350      greece         Greece 2014     1 Full Membership
#> 13     EU   310     hungary        Hungary 2014     1 Full Membership
#> 14     EU   205     ireland        Ireland 2014     1 Full Membership
#> 15     EU   325       italy          Italy 2014     1 Full Membership
#> 16     EU   367      latvia         Latvia 2014     1 Full Membership
#> 17     EU   368   lithuania      Lithuania 2014     1 Full Membership
#> 18     EU   212  luxembourg     Luxembourg 2014     1 Full Membership
#> 19     EU   338       malta          Malta 2014     1 Full Membership
#> 20     EU   210 netherlands    Netherlands 2014     1 Full Membership
#> 21     EU   290      poland         Poland 2014     1 Full Membership
#> 22     EU   235    portugal       Portugal 2014     1 Full Membership
#> 23     EU   360     romania        Romania 2014     1 Full Membership
#> 24     EU   317    slovakia       Slovakia 2014     1 Full Membership
#> 25     EU   349    slovenia       Slovenia 2014     1 Full Membership
#> 26     EU   230       spain          Spain 2014     1 Full Membership
#> 27     EU   380      sweden         Sweden 2014     1 Full Membership
#> 28     EU   200          uk United Kingdom 2014     1 Full Membership
#>           orgname
#> 1  European Union
#> 2  European Union
#> 3  European Union
#> 4  European Union
#> 5  European Union
#> 6  European Union
#> 7  European Union
#> 8  European Union
#> 9  European Union
#> 10 European Union
#> 11 European Union
#> 12 European Union
#> 13 European Union
#> 14 European Union
#> 15 European Union
#> 16 European Union
#> 17 European Union
#> 18 European Union
#> 19 European Union
#> 20 European Union
#> 21 European Union
#> 22 European Union
#> 23 European Union
#> 24 European Union
#> 25 European Union
#> 26 European Union
#> 27 European Union
#> 28 European Union

igo_members("NAFTA", year = c(1995:1998))

#>    ioname ccode  state                 statenme year value        category
#> 1   NAFTA    20 canada                   Canada 1995     1 Full Membership
#> 2   NAFTA    70 mexico                   Mexico 1995     1 Full Membership
#> 3   NAFTA     2    usa United States of America 1995     1 Full Membership
#> 4   NAFTA    20 canada                   Canada 1996     1 Full Membership
#> 5   NAFTA    70 mexico                   Mexico 1996     1 Full Membership
#> 6   NAFTA     2    usa United States of America 1996     1 Full Membership
#> 7   NAFTA    20 canada                   Canada 1997     1 Full Membership
#> 8   NAFTA    70 mexico                   Mexico 1997     1 Full Membership
#> 9   NAFTA     2    usa United States of America 1997     1 Full Membership
#> 10  NAFTA    20 canada                   Canada 1998     1 Full Membership
#> 11  NAFTA    70 mexico                   Mexico 1998     1 Full Membership
#> 12  NAFTA     2    usa United States of America 1998     1 Full Membership
#>               orgname
#> 1  North American FTA
#> 2  North American FTA
#> 3  North American FTA
#> 4  North American FTA
#> 5  North American FTA
#> 6  North American FTA
#> 7  North American FTA
#> 8  North American FTA
#> 9  North American FTA
#> 10 North American FTA
#> 11 North American FTA
#> 12 North American FTA

# Extract different status
igo_members("ACCT", status = c("Associate Membership", "Observer"))

#>    ioname ccode          state               statenme year value
#> 1    ACCT   305        austria                Austria 2014     3
#> 2    ACCT   346         bosnia Bosnia and Herzegovina 2014     3
#> 3    ACCT    94      costarica             Costa Rica 2014     3
#> 4    ACCT   344        croatia                Croatia 2014     3
#> 5    ACCT   352         cyprus                 Cyprus 2014     2
#> 6    ACCT   316       czechrep         Czech Republic 2014     3
#> 7    ACCT    42    domrepublic     Dominican Republic 2014     3
#> 8    ACCT   366        estonia                Estonia 2014     3
#> 9    ACCT   372        georgia                Georgia 2014     3
#> 10   ACCT   310        hungary                Hungary 2014     3
#> 11   ACCT   347         kosovo                 Kosovo 2014     2
#> 12   ACCT   367         latvia                 Latvia 2014     3
#> 13   ACCT   368      lithuania              Lithuania 2014     3
#> 14   ACCT    70         mexico                 Mexico 2014     3
#> 15   ACCT   341     montenegro             Montenegro 2014     3
#> 16   ACCT   541     mozambique             Mozambique 2014     3
#> 17   ACCT   290         poland                 Poland 2014     3
#> 18   ACCT   694          qatar                  Qatar 2014     2
#> 19   ACCT   317       slovakia               Slovakia 2014     3
#> 20   ACCT   349       slovenia               Slovenia 2014     3
#> 21   ACCT   800       thailand               Thailand 2014     3
#> 22   ACCT   696            uae   United Arab Emirates 2014     3
#> 23   ACCT   369        ukraine                Ukraine 2014     3
#> 24   ACCT   165        uruguay                Uruguay 2014     3
#> 25   ACCT   345 yugoslaviaserb             Yugoslavia 2014     3
#>                category            orgname
#> 1              Observer Francophone Agency
#> 2              Observer Francophone Agency
#> 3              Observer Francophone Agency
#> 4              Observer Francophone Agency
#> 5  Associate Membership Francophone Agency
#> 6              Observer Francophone Agency
#> 7              Observer Francophone Agency
#> 8              Observer Francophone Agency
#> 9              Observer Francophone Agency
#> 10             Observer Francophone Agency
#> 11 Associate Membership Francophone Agency
#> 12             Observer Francophone Agency
#> 13             Observer Francophone Agency
#> 14             Observer Francophone Agency
#> 15             Observer Francophone Agency
#> 16             Observer Francophone Agency
#> 17             Observer Francophone Agency
#> 18 Associate Membership Francophone Agency
#> 19             Observer Francophone Agency
#> 20             Observer Francophone Agency
#> 21             Observer Francophone Agency
#> 22             Observer Francophone Agency
#> 23             Observer Francophone Agency
#> 24             Observer Francophone Agency
#> 25             Observer Francophone Agency

# States no members of the UN
igo_members("UN", status = "No Membership")

#>   ioname ccode  state statenme year value      category        orgname
#> 1     UN   347 kosovo   Kosovo 2014     0 No Membership United Nations
#> 2     UN   713 taiwan   Taiwan 2014     0 No Membership United Nations

# Vectorized

igo_members(c("NAFTA","EU"), year = 1993)

#>    ioname ccode       state                 statenme year value        category
#> 1   NAFTA    20      canada                   Canada 1993     1 Full Membership
#> 2   NAFTA    70      mexico                   Mexico 1993     1 Full Membership
#> 3   NAFTA     2         usa United States of America 1993     1 Full Membership
#> 4      EU   211     belgium                  Belgium 1993     1 Full Membership
#> 5      EU   390     denmark                  Denmark 1993     1 Full Membership
#> 6      EU   220      france                   France 1993     1 Full Membership
#> 7      EU   255     germany                  Germany 1993     1 Full Membership
#> 8      EU   350      greece                   Greece 1993     1 Full Membership
#> 9      EU   205     ireland                  Ireland 1993     1 Full Membership
#> 10     EU   325       italy                    Italy 1993     1 Full Membership
#> 11     EU   212  luxembourg               Luxembourg 1993     1 Full Membership
#> 12     EU   210 netherlands              Netherlands 1993     1 Full Membership
#> 13     EU   235    portugal                 Portugal 1993     1 Full Membership
#> 14     EU   230       spain                    Spain 1993     1 Full Membership
#> 15     EU   200          uk           United Kingdom 1993     1 Full Membership
#>               orgname
#> 1  North American FTA
#> 2  North American FTA
#> 3  North American FTA
#> 4      European Union
#> 5      European Union
#> 6      European Union
#> 7      European Union
#> 8      European Union
#> 9      European Union
#> 10     European Union
#> 11     European Union
#> 12     European Union
#> 13     European Union
#> 14     European Union
#> 15     European Union

# Use countrycodes package to get additional codes
if (requireNamespace("countrycode", quietly = TRUE)) {
  library(countrycode)
  EU <- igo_members("EU")
  EU$iso3c <- countrycode(EU$ccode, origin = "cown",
                          destination = "iso3c")
  EU$continent <- countrycode(EU$ccode, origin = "cown",
                              destination = "continent")

  head(EU)
}

#>   ioname ccode    state       statenme year value        category
#> 1     EU   305  austria        Austria 2014     1 Full Membership
#> 2     EU   211  belgium        Belgium 2014     1 Full Membership
#> 3     EU   355 bulgaria       Bulgaria 2014     1 Full Membership
#> 4     EU   344  croatia        Croatia 2014     1 Full Membership
#> 5     EU   352   cyprus         Cyprus 2014     1 Full Membership
#> 6     EU   316 czechrep Czech Republic 2014     1 Full Membership
#>          orgname iso3c continent
#> 1 European Union   AUT    Europe
#> 2 European Union   BEL    Europe
#> 3 European Union   BGR    Europe
#> 4 European Union   HRV    Europe
#> 5 European Union   CYP      Asia
#> 6 European Union   CZE    Europe