Search and Find an IGO

Search any IGO by name or string pattern.

igo_search(pattern = NULL, exact = FALSE)

Arguments

pattern
regex pattern. If NULL the function returns a dataset with all the IGOs on igo_year_format3. Integer values are accepted.

exact
Logical. When TRUE only exact matches are returned.

Value

A dataframe.

Details

The information of each IGO is retrieved based on the last year available on igo_year_format3.

An additional column label is provided. This column is a clean version of longorgname

See also

Examples

# All values
all <- igo_search()

nrow(all)

#> [1] 534

colnames(all)

#>  [1] "ionum"                            "ioname"                          
#>  [3] "orgname"                          "longorgname"                     
#>  [5] "label"                            "sdate"                           
#>  [7] "deaddate"                         "dead"                            
#>  [9] "integrated"                       "replaced"                        
#> [11] "igocode"                          "version"                         
#> [13] "accuracyofpre1965membershipdates" "sourcesandnotes"                 
#> [15] "imputed"                          "political"                       
#> [17] "social"                           "economic"                        

# Search by pattern
igo_search("EU")[,1:3]

#>   ionum      ioname                                              orgname
#> 1    10       ACPEU                                ACP/EU Joint Assembly
#> 2  1830          EU                                       European Union
#> 3  1600       EUFMD             Euro Comm for Control Foot-Mouth Disease
#> 4  1563     EURAMET European Association of National Metrology Institute
#> 5  1565     EURATOM                        European Atomic Energy Agency
#> 6  1730 EUROCONTROL                            Euro Org Safety Air Navig
#> 7  1630    EUROFIMA          Euro Comp for Finance Railway Rolling Stock
#> 8  1590     EUROMET      European Collaboration on Measurement Standards
#> 9  4510         WEU                               Western European Union

igo_search("EU", exact = TRUE)[, 1:3]

#>   ionum ioname        orgname
#> 1  1830     EU European Union

# With integers
igo_search(10)[, 1:3]

#>    ionum   ioname                                               orgname
#> 1    710    AATPO                  Assoc. of Afr. Trade Promotion Orgs.
#> 2     10    ACPEU                                 ACP/EU Joint Assembly
#> 3    100    AFGEC                     Afr. Fund Guarantee & Econ. Coop.
#> 4    410     AFTE                   Arab Federation for Technical Educ.
#> 5    110      AGC                             African Groundnut Council
#> 6    310     AmCC                               Amazonian Coop. Council
#> 7    510     AOMR                       Arab Org. for Mineral Resources
#> 8    610      ARC                               Asian Reinsurance Corp.
#> 9    210      ATO                                   African Timber Org.
#> 10   810      BIS                    Bank for International Settlements
#> 11  3810     BSEC                            Black Sea Economic Council
#> 12  1030    CAECC                Central Asian Cooperation Organization
#> 13  1310   CAMRSD       Conf. African Ministers for Sustain Development
#> 14  1000    CAMSF                          Cent. Am Monetary Stab. Fund
#> 15  1010    CARII      Central American Research Institute for Industry
#> 16  1040      CBI                      Centrl Bureau: Intl Map of World
#> 17  1050     CCNR    Central Commission for the Navigation of the Rhine
#> 18  1060     CCOM                    Cent. Compensation Off. of Maghreb
#> 19  1210      CEC                       Commonwealth Economic Committee
#> 20  1020    CEEPN          Central & Eastern Eur. Privatization Network
#> 21  1070    CEFTA                                    Central Europe FTA
#> 22  1080      CEI                           Central European Initiative
#> 23  1100    CENTO                           Central Treaty Organization
#> 24  1510    CEPGL              Economic Community of Great Lakes States
#> 25   910    CFATF                      Caribbean Fin. Action Task Force
#> 26  1410   CMHASG         Council of Min. of Health of Arab Gulf States
#> 27  1095     CPAB                         Central PanAm Bureau Eugenics
#> 28  1610     ECCD                       Euro Comm for Control of Danube
#> 29  1810     ELDO                   Euro Space Vehicle Launcher Dev Org
#> 30  1710     EMBL                              Eur. Molecular Biol. Lab
#> 31  1910      G15                                           Group of 15
#> 32  2010    HCPIL                        Hague Conf on Private Intl Law
#> 33  2110     IACI                            Inter-Am Children's Instit
#> 34  2100     IARA                        Inter Allied Reparation Agency
#> 35  2105    IARHC                      Inter Allied Rhineland High Comm
#> 36  2210    IATTC                           Inter-Am Tropical Tuna Comm
#> 37  2410      IBA                                  Int'l Bauxite Assoc.
#> 38  2610   ICCSLT                       Intl Comm of Cape Spartel Light
#> 39  2510   ICivDO                              Int'l Civil Defence Org.
#> 40  2710      IEA                                    Intl Energy Agency
#> 41  2810      IIF                          Intl Instit of Refrigeration
#> 42  1110   INFSMK                                             InfoSAMAK
#> 43  2910     INSG                              Int'l Nickel Study Group
#> 44  3010    IPhyL                                Intl Physiological Lab
#> 45  3100      ISA                                Int'l Seabed Authority
#> 46  3310  ISHREST     Inter-State School for Hydraulic & Rural Engin...
#> 47  3110    ISUPT                     Intl Sec for Unification Pharmaco
#> 48  2310    ITVRC                                Intgvt TV & Radio Corp
#> 49  3210   IUPLAW          Intl Union for Protection of Lit & Art Works
#> 50  3410     LAEO                            Latin American Energy Org.
#> 51  3510  MCWCASM Minist Conf of West & Cent African States on Maritime
#> 52  3610     NCTR                       Nordic Council for Tax Research
#> 53  3710    NPFSC                           North Pacific Fur Seal Comm
#> 54  1090     OTIF             Central Office for Intl Railway Transport
#> 55  3910       PC                                      Paris Commission
#> 56  4010   PMAESA            Port Management Assoc. of E. and S. Africa
#> 57  4100 RIOgroup                                             Rio Group
#> 58  4110     SCAf                             Scientific Council in SSA
#> 59  4210    SEATO                    Southeast Asia Treaty Organization
#> 60  4410   UNESCO              UN Education, Scientific, & Cultural Org
#> 61  4450    VASAB                 Vision & Strategies Around Baltic Sea
#> 62  4510      WEU                                Western European Union

igo_search(10, exact = TRUE)[, 1:3]

#>   ionum ioname               orgname
#> 1    10  ACPEU ACP/EU Joint Assembly

# Several patterns (regex style)
igo_search("NAFTA|UN|EU")[,1:3]

#>    ionum      ioname                                              orgname
#> 1     10       ACPEU                                ACP/EU Joint Assembly
#> 2    430     AGPUNDO                     Arab Gulf Prog. for UN Dev. Org.
#> 3   1830          EU                                       European Union
#> 4   1600       EUFMD             Euro Comm for Control Foot-Mouth Disease
#> 5   1563     EURAMET European Association of National Metrology Institute
#> 6   1565     EURATOM                        European Atomic Energy Agency
#> 7   1730 EUROCONTROL                            Euro Org Safety Air Navig
#> 8   1630    EUROFIMA          Euro Comp for Finance Railway Rolling Stock
#> 9   1590     EUROMET      European Collaboration on Measurement Standards
#> 10  3670       NAFTA                                   North American FTA
#> 11  4400          UN                                       United Nations
#> 12  4410      UNESCO             UN Education, Scientific, & Cultural Org
#> 13  4420       UNIDO                        UN Industrial Development Org
#> 14  2790    UNIDROIT           Intl Instit for Unification of Private Law
#> 15  4510         WEU                               Western European Union

# Several patterns Exact (regex style)
igo_search("^NAFTA$|^UN$|^EU$")[,1:3]

#>   ionum ioname            orgname
#> 1  1830     EU     European Union
#> 2  3670  NAFTA North American FTA
#> 3  4400     UN     United Nations