Search and Find an IGO

Search any IGO by name or string pattern.

igo_search(pattern = NULL, exact = FALSE)

Arguments

pattern
regex pattern. If NULL the function returns a dataset with all the IGOs on igo_year_format3. Integer values are accepted.

exact
Logical. When TRUE only exact matches are returned.

Value

A dataframe.

Details

The information of each IGO is retrieved based on the last year available on igo_year_format3.

An additional column label is provided. This column is a clean version of longorgname

See also

Examples

# All values
all <- igo_search()

nrow(all)

#> [1] 534

colnames(all)

#>  [1] "ionum"                            "ioname"                          
#>  [3] "orgname"                          "longorgname"                     
#>  [5] "label"                            "sdate"                           
#>  [7] "deaddate"                         "dead"                            
#>  [9] "integrated"                       "replaced"                        
#> [11] "igocode"                          "version"                         
#> [13] "accuracyofpre1965membershipdates" "sourcesandnotes"                 
#> [15] "imputed"                          "political"                       
#> [17] "social"                           "economic"                        

# Search by pattern
igo_search("EU")[,1:3]

#>    ionum      ioname                                               orgname
#> 1    680    ABEPSEAC      Assoc. B/t EEC and States of East Afr. Community
#> 2     10       ACPEU                                 ACP/EU Joint Assembly
#> 3    150        APPA                    African Petroleum Producers Assoc.
#> 4    540        ASEF                                Asia-Europe Foundation
#> 5   1020       CEEPN          Central & Eastern Eur. Privatization Network
#> 6   1070       CEFTA                                    Central Europe FTA
#> 7   1080         CEI                           Central European Initiative
#> 8   1720        CERN                     European Org for Nuclear Research
#> 9   1390         COE                                     Council of Europe
#> 10  1300    CONFEJES          Conference des Ministres jeunesse...Francais
#> 11  1095        CPAB                         Central PanAm Bureau Eugenics
#> 12  1545        EAEC                           Eurasian Economic Community
#> 13  1550        EAPC                     Euro-Atlantic Partnership Council
#> 14  1540        EAPO                                  Eurasian Patent Org.
#> 15  1570        EBRD        European Bank for Reconstruction & Development
#> 16  1580         ECB                                 European Central Bank
#> 17  1610        ECCD                       Euro Comm for Control of Danube
#> 18  1620      ECCPIF                 Euro Company Chem Process Irrad Fuels
#> 19  1640       ECPTA                        Euro Conf Postal Telecom Admin
#> 20  1585        ECSC                       European Coal & Steel Community
#> 21  1653         EEC                           European Economic Community
#> 22  1645        EFCC                                Euro Food Code Council
#> 23  1660      EFILWC Eur. Foundation for Impr. of Living & Work Conditions
#> 24  1670        EFTA                                  Euro Free Trade Assn
#> 25  1690         EIB                                  Euro Investment Bank
#> 26  1675         EIP                           European Institute of Peace
#> 27  1680        EIPA                     Europ. Institute of Public Admin.
#> 28  1810        ELDO                   Euro Space Vehicle Launcher Dev Org
#> 29  1700        EMBC                 European Molecular Biology Conference
#> 30  1710        EMBL                              Eur. Molecular Biol. Lab
#> 31  1715         EMI                           European Monetary Institute
#> 32  1560       EMPPO                          Euro & Med Plant Protect Org
#> 33  1770         EPA                              Euro Productivity Agency
#> 34  1760       EPFSC                 Euro. Postal Financial Services Comm.
#> 35  1740         EPO                          European Patent Organization
#> 36  1750         EPU                               European Payments Union
#> 37  1790         ESA                                 European Space Agency
#> 38  1780         ESO                         European Southern Observatory
#> 39  1800        ESRO                               Euro Space Research Org
#> 40  1820         ETF                          European Training Foundation
#> 41  1830          EU                                        European Union
#> 42  1600       EUFMD              Euro Comm for Control Foot-Mouth Disease
#> 43  1563     EURAMET  European Association of National Metrology Institute
#> 44  1565     EURATOM                         European Atomic Energy Agency
#> 45  1730 EUROCONTROL                             Euro Org Safety Air Navig
#> 46  1630    EUROFIMA           Euro Comp for Finance Railway Rolling Stock
#> 47  1590     EUROMET       European Collaboration on Measurement Standards
#> 48  2250       IOMig                                Intl Org for Migration
#> 49  3800       OAPEC                 Org. Arab Petroleum Export. Countries
#> 50  3762        OEEC                        Org for European Economic Coop
#> 51  3840        OPEC                  Org of Petroleum Exporting Countries
#> 52  3770        OSCE                      Org. Security Cooperation Europe
#> 53  4000         PED                          Pole European de Development
#> 54  1650         WCO                                     World Customs Org
#> 55  4510         WEU                                Western European Union

igo_search("EU", exact = TRUE)[, 1:3]

#>   ionum ioname        orgname
#> 1  1830     EU European Union

# With integers
igo_search(10)[, 1:3]

#>    ionum   ioname                                               orgname
#> 1    710    AATPO                  Assoc. of Afr. Trade Promotion Orgs.
#> 2     10    ACPEU                                 ACP/EU Joint Assembly
#> 3    100    AFGEC                     Afr. Fund Guarantee & Econ. Coop.
#> 4    410     AFTE                   Arab Federation for Technical Educ.
#> 5    110      AGC                             African Groundnut Council
#> 6    310     AmCC                               Amazonian Coop. Council
#> 7    510     AOMR                       Arab Org. for Mineral Resources
#> 8    610      ARC                               Asian Reinsurance Corp.
#> 9    210      ATO                                   African Timber Org.
#> 10   810      BIS                    Bank for International Settlements
#> 11  3810     BSEC                            Black Sea Economic Council
#> 12  1030    CAECC                Central Asian Cooperation Organization
#> 13  1310   CAMRSD       Conf. African Ministers for Sustain Development
#> 14  1000    CAMSF                          Cent. Am Monetary Stab. Fund
#> 15  1010    CARII      Central American Research Institute for Industry
#> 16  1040      CBI                      Centrl Bureau: Intl Map of World
#> 17  1050     CCNR    Central Commission for the Navigation of the Rhine
#> 18  1060     CCOM                    Cent. Compensation Off. of Maghreb
#> 19  1210      CEC                       Commonwealth Economic Committee
#> 20  1020    CEEPN          Central & Eastern Eur. Privatization Network
#> 21  1070    CEFTA                                    Central Europe FTA
#> 22  1080      CEI                           Central European Initiative
#> 23  1100    CENTO                           Central Treaty Organization
#> 24  1510    CEPGL              Economic Community of Great Lakes States
#> 25   910    CFATF                      Caribbean Fin. Action Task Force
#> 26  1410   CMHASG         Council of Min. of Health of Arab Gulf States
#> 27  1095     CPAB                         Central PanAm Bureau Eugenics
#> 28  1610     ECCD                       Euro Comm for Control of Danube
#> 29  1810     ELDO                   Euro Space Vehicle Launcher Dev Org
#> 30  1710     EMBL                              Eur. Molecular Biol. Lab
#> 31  1910      G15                                           Group of 15
#> 32  2010    HCPIL                        Hague Conf on Private Intl Law
#> 33  2110     IACI                            Inter-Am Children's Instit
#> 34  2100     IARA                        Inter Allied Reparation Agency
#> 35  2105    IARHC                      Inter Allied Rhineland High Comm
#> 36  2210    IATTC                           Inter-Am Tropical Tuna Comm
#> 37  2410      IBA                                  Int'l Bauxite Assoc.
#> 38  2610   ICCSLT                       Intl Comm of Cape Spartel Light
#> 39  2510   ICivDO                              Int'l Civil Defence Org.
#> 40  2710      IEA                                    Intl Energy Agency
#> 41  2810      IIF                          Intl Instit of Refrigeration
#> 42  1110   INFSMK                                             InfoSAMAK
#> 43  2910     INSG                              Int'l Nickel Study Group
#> 44  3010    IPhyL                                Intl Physiological Lab
#> 45  3100      ISA                                Int'l Seabed Authority
#> 46  3310  ISHREST     Inter-State School for Hydraulic & Rural Engin...
#> 47  3110    ISUPT                     Intl Sec for Unification Pharmaco
#> 48  2310    ITVRC                                Intgvt TV & Radio Corp
#> 49  3210   IUPLAW          Intl Union for Protection of Lit & Art Works
#> 50  3410     LAEO                            Latin American Energy Org.
#> 51  3510  MCWCASM Minist Conf of West & Cent African States on Maritime
#> 52  3610     NCTR                       Nordic Council for Tax Research
#> 53  3710    NPFSC                           North Pacific Fur Seal Comm
#> 54  1090     OTIF             Central Office for Intl Railway Transport
#> 55  3910       PC                                      Paris Commission
#> 56  4010   PMAESA            Port Management Assoc. of E. and S. Africa
#> 57  4100 RIOgroup                                             Rio Group
#> 58  4110     SCAf                             Scientific Council in SSA
#> 59  4210    SEATO                    Southeast Asia Treaty Organization
#> 60  4410   UNESCO              UN Education, Scientific, & Cultural Org
#> 61  4450    VASAB                 Vision & Strategies Around Baltic Sea
#> 62  4510      WEU                                Western European Union

igo_search(10, exact = TRUE)[, 1:3]

#>   ionum ioname               orgname
#> 1    10  ACPEU ACP/EU Joint Assembly

# Several patterns (regex style)
igo_search("NAFTA|UN|EU")[,1:3]

#>     ionum      ioname
#> 1     680    ABEPSEAC
#> 2     400         ACC
#> 3      10       ACPEU
#> 4     570         ACU
#> 5     420       AFESD
#> 6     100       AFGEC
#> 7     160        AFPU
#> 8      90      AFRAND
#> 9     725       AFSPC
#> 10    110         AGC
#> 11    430     AGPUNDO
#> 12    590        AIDC
#> 13    730       AIOEC
#> 14    310        AmCC
#> 15    125       AMCOW
#> 16    490         AMF
#> 17    250       AMPTU
#> 18     40        AMSC
#> 19    470         AMU
#> 20    330      Andean
#> 21    360       ANZUS
#> 22    640        AOPU
#> 23    550        APCC
#> 24   2060       APFIC
#> 25    150        APPA
#> 26    670         APT
#> 27    155        APTU
#> 28   2770     AralSea
#> 29    520        ARPU
#> 30    530      ArticC
#> 31    740   ASCRubber
#> 32    540        ASEF
#> 33    560       ASPAC
#> 34    770        ATPC
#> 35   3761          AU
#> 36    780          BC
#> 37    870        BCSC
#> 38    840     BENELUX
#> 39   3810        BSEC
#> 40   1180       CAARC
#> 41   1030       CAECC
#> 42     30       CAMES
#> 43   1000       CAMSF
#> 44    880     CARICOM
#> 45   1200        CATC
#> 46   1420        CBSS
#> 47   4485        CEAO
#> 48   1020       CEEPN
#> 49   1070       CEFTA
#> 50   1080         CEI
#> 51   1260       CEMAC
#> 52   1510       CEPGL
#> 53   1720        CERN
#> 54   1160         CFC
#> 55   1400       CMAEC
#> 56   1370        CMEA
#> 57   1410      CMHASG
#> 58   1390         COE
#> 59   1380     COLOMBO
#> 60   1300    CONFEJES
#> 61   1340      COPTAC
#> 62   1095        CPAB
#> 63   1270        CPSC
#> 64    920         CPU
#> 65   1250         CTO
#> 66    900         CXC
#> 67   4120         EAC
#> 68   1545        EAEC
#> 69   1550        EAPC
#> 70   1540        EAPO
#> 71   1570        EBRD
#> 72   1580         ECB
#> 73   1500       ECCAS
#> 74   1610        ECCD
#> 75   1620      ECCPIF
#> 76   1520      ECOWAS
#> 77   1640       ECPTA
#> 78   1585        ECSC
#> 79   1653         EEC
#> 80   1645        EFCC
#> 81   1660      EFILWC
#> 82   1670        EFTA
#> 83   1690         EIB
#> 84   1675         EIP
#> 85   1680        EIPA
#> 86   1810        ELDO
#> 87   1700        EMBC
#> 88   1710        EMBL
#> 89   1715         EMI
#> 90   1560       EMPPO
#> 91   1430     Entente
#> 92   1770         EPA
#> 93   1760       EPFSC
#> 94   1740         EPO
#> 95   1750         EPU
#> 96   1790         ESA
#> 97   1780         ESO
#> 98   1800        ESRO
#> 99   1820         ETF
#> 100  1830          EU
#> 101  1600       EUFMD
#> 102  1563     EURAMET
#> 103  1565     EURATOM
#> 104  1730 EUROCONTROL
#> 105  1630    EUROFIMA
#> 106  1590     EUROMET
#> 107  1860     FDIPLAC
#> 108  1990         GCC
#> 109  1980      GCRSNC
#> 110  1900         GEF
#> 111  1920     GLACSEC
#> 112  2340       IATSJ
#> 113  2210       IATTC
#> 114  2280       ICCEC
#> 115  2690        ICES
#> 116  2650       ICPTU
#> 117  2260      ICRPBC
#> 118  2760        IFAD
#> 119  3260         IGC
#> 120  2880         IMF
#> 121  3170    INTELSAT
#> 122  2250       IOMig
#> 123  2960        IOOC
#> 124  2950       IOPCF
#> 125  3000         IPC
#> 126  3070         IRU
#> 127  3130        ISuC
#> 128  3110       ISUPT
#> 129  3180         ITC
#> 130  3160         ITU
#> 131  2315        IUIC
#> 132  3220       IUPCT
#> 133  3200       IUPIP
#> 134  3210      IUPLAW
#> 135  4360      IUPNVP
#> 136  3230        IUPR
#> 137  3400       LAIEC
#> 138  3440       LATIN
#> 139  3520      MCPTTC
#> 140  3530    Montreal
#> 141  3480         MRU
#> 142  3670       NAFTA
#> 143  3620         NCM
#> 144  3600        NCRR
#> 145  3610        NCTR
#> 146  3630         NDF
#> 147  3640        NERC
#> 148  3590       NordC
#> 149  3660        NTSC
#> 150  3665        NWHF
#> 151  3800       OAPEC
#> 152  3900         OAS
#> 153  3760         OAU
#> 154   260        OCAM
#> 155  3762        OEEC
#> 156  3840        OPEC
#> 157  3770        OSCE
#> 158  3905        PAPU
#> 159  4000         PED
#> 160  4020       PUASP
#> 161  4040      RadioU
#> 162  4050      RASCOM
#> 163  4080      RCAELA
#> 164  4075         RCC
#> 165  4060        RCFC
#> 166  4270       SAAFA
#> 167  4240        SACU
#> 168  4250        SADC
#> 169  4230       SARTC
#> 170  4110        SCAf
#> 171  4300         SCH
#> 172  4290        SugU
#> 173  4150       SWAPU
#> 174  4380        UASC
#> 175  4370        UBEC
#> 176   940       UDEAC
#> 177  4340       UEMOA
#> 178  4350       UIUCV
#> 179  4390       UKDWD
#> 180  4375          UM
#> 181  4365        UMAC
#> 182  4503        UMOA
#> 183  4400          UN
#> 184  4410      UNESCO
#> 185  4420       UNIDO
#> 186  2790    UNIDROIT
#> 187  4430         UPU
#> 188  4440         USP
#> 189  1940    VALDIVIA
#> 190  4450       VASAB
#> 191  4490        WAEC
#> 192  4500        WAHC
#> 193  4540        WCDC
#> 194  1650         WCO
#> 195  4510         WEU
#> 196  4520         WNF
#>                                                             orgname
#> 1                  Assoc. B/t EEC and States of East Afr. Community
#> 2                                          Arab Cooperation Council
#> 3                                             ACP/EU Joint Assembly
#> 4                                              Asian Clearing Union
#> 5                         Arab Fund for Social/Economic Development
#> 6                                 Afr. Fund Guarantee & Econ. Coop.
#> 7                                              African Postal Union
#> 8                                         Afr. Foundation for R & D
#> 9         Association of Financial Supervisors of Pacific Countries
#> 10                                        African Groundnut Council
#> 11                                 Arab Gulf Prog. for UN Dev. Org.
#> 12                                Asian Industrial Develop. Council
#> 13                                  Assoc. Iron Ore Expt. Countries
#> 14                                          Amazonian Coop. Council
#> 15                              African Ministers' Council on Water
#> 16                                               Arab Monetary Fund
#> 17                          Afro-Malagasy Postal and Telecomm Union
#> 18                                    African/Malgasy Sugar Council
#> 19                                               Arab Maghreb Union
#> 20                                                 Andean Community
#> 21                      Australia-NewZealand-US Treaty Organization
#> 22                                       Asian-Oceanic Postal Union
#> 23                                      Asian/Pacific Coconut Comm.
#> 24                                      Asia Pacific Fisheries Comm
#> 25                               African Petroleum Producers Assoc.
#> 26                                      Asian-Pacific Telecommunity
#> 27                                 African Postal and Telecom Union
#> 28                                   Int'l Fund for Saving Aral Sea
#> 29                                                Arab Postal Union
#> 30                                                   Arctic Council
#> 31                           Association of Rubber Producing States
#> 32                                           Asia-Europe Foundation
#> 33                                           Asia & Pacific Council
#> 34                                   Assoc. Tin Producing Countries
#> 35                                                    African Union
#> 36                                                   Baltic Council
#> 37                             British Commonwealth Scientific Comm
#> 38                                                Benelux Community
#> 39                                       Black Sea Economic Council
#> 40                       Commonwealth Advis. Aero. Research Council
#> 41                           Central Asian Cooperation Organization
#> 42                           Afr. & Malagasy Council for Higher Ed.
#> 43                                     Cent. Am Monetary Stab. Fund
#> 44                                              Caribbean Community
#> 45                               Commonwealth Air Transport Council
#> 46                                     Council of Baltic Sea States
#> 47                                  West African Economic Community
#> 48                     Central & Eastern Eur. Privatization Network
#> 49                                               Central Europe FTA
#> 50                                      Central European Initiative
#> 51                        Central African Economic & Monetary Union
#> 52                         Economic Community of Great Lakes States
#> 53                                European Org for Nuclear Research
#> 54                                      Common Fund for Commodities
#> 55                                Council of Min. Asian Econ. Coop.
#> 56                                  Council for Mutual Economic Aid
#> 57                    Council of Min. of Health of Arab Gulf States
#> 58                                                Council of Europe
#> 59                             Council for Tech Coop in S & SE Asia
#> 60                     Conference des Ministres jeunesse...Francais
#> 61           Conf. of Posts & Telecommunications  of Central Africa
#> 62                                    Central PanAm Bureau Eugenics
#> 63                                Comm Portugese Speaking Countries
#> 64                                           Caribbean Postal Union
#> 65                                       Commonwealth Telecom Board
#> 66                                     Caribb. Examinations Council
#> 67                                           East African Community
#> 68                                      Eurasian Economic Community
#> 69                                Euro-Atlantic Partnership Council
#> 70                                             Eurasian Patent Org.
#> 71                   European Bank for Reconstruction & Development
#> 72                                            European Central Bank
#> 73                     Economic Community of Central African States
#> 74                                  Euro Comm for Control of Danube
#> 75                            Euro Company Chem Process Irrad Fuels
#> 76                        Economic Community of West African States
#> 77                                   Euro Conf Postal Telecom Admin
#> 78                                  European Coal & Steel Community
#> 79                                      European Economic Community
#> 80                                           Euro Food Code Council
#> 81            Eur. Foundation for Impr. of Living & Work Conditions
#> 82                                             Euro Free Trade Assn
#> 83                                             Euro Investment Bank
#> 84                                      European Institute of Peace
#> 85                                Europ. Institute of Public Admin.
#> 86                              Euro Space Vehicle Launcher Dev Org
#> 87                            European Molecular Biology Conference
#> 88                                         Eur. Molecular Biol. Lab
#> 89                                      European Monetary Institute
#> 90                                     Euro & Med Plant Protect Org
#> 91                                                  Entente Council
#> 92                                         Euro Productivity Agency
#> 93                            Euro. Postal Financial Services Comm.
#> 94                                     European Patent Organization
#> 95                                          European Payments Union
#> 96                                            European Space Agency
#> 97                                    European Southern Observatory
#> 98                                          Euro Space Research Org
#> 99                                     European Training Foundation
#> 100                                                  European Union
#> 101                        Euro Comm for Control Foot-Mouth Disease
#> 102            European Association of National Metrology Institute
#> 103                                   European Atomic Energy Agency
#> 104                                       Euro Org Safety Air Navig
#> 105                     Euro Comp for Finance Railway Rolling Stock
#> 106                 European Collaboration on Measurement Standards
#> 107                    Fund 4 Devel. of Indigen Peoples of LA/Carib
#> 108                                        Gulf Cooperation Council
#> 109                  Guidance Comm. for Road Safety in Nordic Cntry
#> 110                                            Global Environ. Fund
#> 111                      Group of L/A & Carib. Sugar Exp. Countries
#> 112                               Intl Arbitration Trib at San Jose
#> 113                                     Inter-Am Tropical Tuna Comm
#> 114                         Intgvt Council of Copper Exp. Countries
#> 115                             Intl Council for Exploration of Sea
#> 116                            Intl Conf Promoting Tech Unification
#> 117                    Intergov. Committee of the River Plate Basin
#> 118                          Int Fund for Agriculture & Development
#> 119                                             Intl Grains Council
#> 120                                              Intl Monetary Fund
#> 121                                  Int'l Telecomm. Satellite Org.
#> 122                                          Intl Org for Migration
#> 123                                          Intl Olive Oil Council
#> 124                              Int'l Oil Pollution Compens. Funds
#> 125                                          Int'l Pepper Community
#> 126                                               Intl Relief Union
#> 127                                              Intl Sugar Council
#> 128                               Intl Sec for Unification Pharmaco
#> 129                                                Intl Tin Council
#> 130                                              Intl Telecom Union
#> 131                     Intergovernmental University of Cooperation
#> 132                       Intl Union Publication of Customs Tariffs
#> 133 Intl Union for Protection of Industrial Prop (Paris Convention)
#> 134                    Intl Union for Protection of Lit & Art Works
#> 135                 Union for Protection of New Varieties of Plants
#> 136                                             Intl Union of Pruth
#> 137           Latin American Institute of Educational Communication
#> 138                                                     Latin Union
#> 139                 Multi-Country Posts & Telecomm. Training Center
#> 140      Multilat. Fund for the Implement. of the Montreal Protocol
#> 141                                                Mano River Union
#> 142                                              North American FTA
#> 143                                     Nordic Council of Ministers
#> 144                                 Nord. Council for Reindeer Res.
#> 145                                 Nordic Council for Tax Research
#> 146                                         Nordic Development Fund
#> 147                                        Nordic Econ Res. Council
#> 148                                                  Nordic Council
#> 149                               Nordic Telcomm. Satellite Council
#> 150                                Nordic World Heritage Foundation
#> 151                           Org. Arab Petroleum Export. Countries
#> 152                                   Pan-Am Union/Org of Am States
#> 153                                           Org for African Unity
#> 154                               Common Afro-Malagasy Economic Org
#> 155                                  Org for European Economic Coop
#> 156                            Org of Petroleum Exporting Countries
#> 157                                Org. Security Cooperation Europe
#> 158                                        Pan African Postal Union
#> 159                                    Pole European de Development
#> 160                                 Postal Union of Am, Spain, Port
#> 161                                            Radiotelegraph Union
#> 162                                  Reg. Afr. Satellite Comm. Org.
#> 163               Reg. Council for Adult Educ. & Literacy in Africa
#> 164                                    Regional Cooperation Council
#> 165                           Reg Commonwealth in the Field of Comm
#> 166                                Special Arab Aid Fund for Africa
#> 167                                             Southern African CU
#> 168                                 Southern African Dev. Community
#> 169                          South African Regional Tourism Council
#> 170                                       Scientific Council in SSA
#> 171                                      Superior Council of Health
#> 172                                                     Sugar Union
#> 173                                  South & West Asia Postal Union
#> 174                                        United Arab Shipping Co.
#> 175                             Union of Banana Exporting Countries
#> 176                        Central African Customs & Economic Union
#> 177                          West African Economic & Monetary Union
#> 178                          Union for Intl Use of Carriages & Vans
#> 179                                      UK Dominion Wool Disposals
#> 180                                     Union for the Mediterranean
#> 181                                  Central African Monetary Union
#> 182                                     West African Monetary Union
#> 183                                                  United Nations
#> 184                        UN Education, Scientific, & Cultural Org
#> 185                                   UN Industrial Development Org
#> 186                      Intl Instit for Unification of Private Law
#> 187                                          Universal Postal Union
#> 188                                          Univ. of South Pacific
#> 189                           Group of S. Hemisp. Countries on Env.
#> 190                           Vision & Strategies Around Baltic Sea
#> 191                                  West Afr. Examinations Council
#> 192                                   West African Health Community
#> 193                           Working Community of Danube Countries
#> 194                                               World Customs Org
#> 195                                          Western European Union
#> 196                                                West Nordic Fund

# Several patterns Exact (regex style)
igo_search("^NAFTA$|^UN$|^EU$")[,1:3]

#>   ionum ioname            orgname
#> 1  1830     EU     European Union
#> 2  3670  NAFTA North American FTA
#> 3  4400     UN     United Nations