Intergovernmental Organizations (IGO) by year

Data on IGOs from 1815-2014, at the IGO-year level. Contains one record per IGO-year (with years listed at 5 year intervals through 1965, and annually thereafter).

Format

data.frame with 19,335 rows. Relevant fields:

ioname: Short abbreviation of the IGO name.

orgname: Full IGO name.

year: Calendar Year.

afghanistan...zimbabwe: status of that state in the IGO. See Details.

sdate: start date (year) that the IGO started.

deaddate: dead date (year) that the IGO ceased to exist.

longorgname: a longer version of the IGOs name (including previous names)

ionum: IGO id number in v2.1 and v3.0 of the data.

version: COW version number.

See Codebook Version 3 IGO Data for full reference.

Source

Intergovernmental Organizations (v3), The Correlates of War Project (IGO Data Stata Files).

Details

Possible values of the status of that state in the IGO are:

Category	Numerical Value

No Membership	0

Full Membership	1

Associate Membership	2

Observer	3

Missing data	-9

State Not System Member	-1

See igo_recode_igoyear() section for an easy way to recode the numerical values into factors.

Note

Raw data used internally by igoR.

References

Pevehouse, J. C., Nordstrom, T., McManus, R. W., & Jamison, A. S. (2020). Tracking organizations in the world: The Correlates of War IGO Version 3.0 datasets. Journal of Peace Research, 57(3), 492–503. doi:10.1177/0022343319881175 .

See also

Examples

data("state_year_format3")

# Show a glimpse
library(dplyr)

state_year_format3 %>%
  select(ccode:afgec) %>%
  filter(year > 1990) %>%
  glimpse()
#> Rows: 4,567
#> Columns: 26
#> $ ccode    <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2…
#> $ year     <dbl> 1991, 1992, 1993, 1994, 1995, 1996, 1997, 1998, 1999, 2000, 2…
#> $ state    <chr> "usa", "usa", "usa", "usa", "usa", "usa", "usa", "usa", "usa"…
#> $ aaaid    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aacb     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aalco    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aaro     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aata     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aatpo    <dbl> 0, 0, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,…
#> $ abeda    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abepseac <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acc      <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acct     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acdt     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ aci      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -1, -1, -1, -1, -1, -1, -1, -1…
#> $ acml     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -1, -1, -1, -1, -1, -1, -1,…
#> $ acp      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acpeu    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acs      <dbl> -1, -1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acso     <dbl> -9, -9, -9, -9, -9, -9, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acssrb   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acu      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acwl     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 0, 0, 0, 0, 0, 0, 0, …
#> $ afesd    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ afeximb  <dbl> -1, -1, -9, -9, -9, -9, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, -…
#> $ afgec    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…

# Recode numerical to factors: with a sample
sample_state_year <- state_year_format3 %>%
  as_tibble() %>%
  select(ccode:afgec) %>%
  filter(year == 1990)

sample_state_year %>% glimpse()
#> Rows: 161
#> Columns: 26
#> $ ccode    <dbl> 2, 20, 31, 40, 41, 42, 51, 52, 53, 54, 55, 56, 57, 58, 60, 70…
#> $ year     <dbl> 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1…
#> $ state    <chr> "usa", "canada", "bahamas", "cuba", "haiti", "domrepublic", "…
#> $ aaaid    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aacb     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aalco    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aaro     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aata     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aatpo    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abeda    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abepseac <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acc      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acct     <dbl> 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acdt     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ aci      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acml     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acp      <dbl> 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0…
#> $ acpeu    <dbl> 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0…
#> $ acs      <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acso     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acssrb   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acu      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acwl     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ afesd    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ afeximb  <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ afgec    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…

# Recode
sample_state_year_recoded <- sample_state_year %>%
  mutate(across(-c(ccode:state), igo_recode_stateyear))

sample_state_year_recoded %>% glimpse()
#> Rows: 161
#> Columns: 26
#> $ ccode    <dbl> 2, 20, 31, 40, 41, 42, 51, 52, 53, 54, 55, 56, 57, 58, 60, 70…
#> $ year     <dbl> 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1…
#> $ state    <chr> "usa", "canada", "bahamas", "cuba", "haiti", "domrepublic", "…
#> $ aaaid    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aacb     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aalco    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aaro     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aata     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aatpo    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ abeda    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ abepseac <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acc      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acct     <fct> No Membership, Full Membership, No Membership, No Membership,…
#> $ acdt     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ aci      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acml     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acp      <fct> No Membership, No Membership, Full Membership, No Membership,…
#> $ acpeu    <fct> No Membership, No Membership, Full Membership, No Membership,…
#> $ acs      <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acso     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acssrb   <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acu      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acwl     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ afesd    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ afeximb  <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ afgec    <fct> No Membership, No Membership, No Membership, No Membership, N…