Intergovernmental Organizations (IGO) by year

Data on IGOs from 1815-2014, at the IGO-year level. Contains one record per IGO-year (with years listed at 5 year intervals through 1965, and annually thereafter).

Format

data.frame with 19,335 rows. Relevant fields:

ioname: Short abbreviation of the IGO name.

orgname: Full IGO name.

year: Calendar Year.

afghanistan...zimbabwe: status of that state in the IGO. See Details.

sdate: start date (year) that the IGO started.

deaddate: dead date (year) that the IGO dead.

longorgname: a longer version of the IGOs name (including previous names)

ionum: IGO id number in v2.1 and v3.0 of the data.

version: COW version number.

See Codebook Version 3 IGO Data for full reference.

Source

Intergovernmental Organizations (v3), The Correlates of War Project (IGO Data Stata Files).

Details

Possible value of the status of that state in the IGO are:

Category	Numerical Value

No Membership	0

Full Membership	1

Associate Membership	2

Observer	3

Missing data	-9

State Not System Member	-1

See igo_recode_igoyear() section for an easy way to recode the numerical values into factors.

Note

Raw data used internally by igoR.

References

Pevehouse, J. C., Nordstrom, T., McManus, R. W., & Jamison, A. S. (2020). Tracking organizations in the world: The Correlates of War IGO Version 3.0 datasets. Journal of Peace Research, 57(3), 492–503. doi:10.1177/0022343319881175 .

See also

Examples

data("state_year_format3")

# Show a glimpse
library(dplyr)
names(state_year_format3)
#>   [1] "ccode"       "year"        "state"       "aaaid"       "aacb"       
#>   [6] "aalco"       "aaro"        "aata"        "aatpo"       "abeda"      
#>  [11] "abepseac"    "acc"         "acct"        "acdt"        "aci"        
#>  [16] "acml"        "acp"         "acpeu"       "acs"         "acso"       
#>  [21] "acssrb"      "acu"         "acwl"        "afesd"       "afeximb"    
#>  [26] "afgec"       "afpu"        "afrand"      "afristat"    "afspc"      
#>  [31] "afte"        "agc"         "agpundo"     "aic"         "aidc"       
#>  [36] "aido"        "aioec"       "aipo"        "aitic"       "alo"        
#>  [41] "alsf"        "amco"        "amcow"       "amf"         "amipo"      
#>  [46] "amptu"       "amsc"        "amu"         "anzus"       "aoad"       
#>  [51] "aocrs"       "aomr"        "aopu"        "ap"          "apcc"       
#>  [56] "apec"        "apfic"       "apibd"       "apo"         "appa"       
#>  [61] "apt"         "aptu"        "arc"         "arcal"       "aripo"      
#>  [66] "arpu"        "asatp"       "asblac"      "ascbc"       "ascrubber"  
#>  [71] "asean"       "asecna"      "asef"        "aspac"       "ato"        
#>  [76] "atpc"        "au"          "avrdc"       "afdb"        "afropda"    
#>  [81] "africarice"  "africare"    "amcc"        "andean"      "aralsea"    
#>  [86] "articc"      "asdb"        "baltbat"     "bc"          "bcsc"       
#>  [91] "benelux"     "biisef"      "bionet"      "bipm"        "bis"        
#>  [96] "bndp"        "bobp"        "bonn"        "borgip"      "bs"         
#> [101] "bsec"        "caad"        "caarc"       "cab"         "cabi"       
#> [106] "cacb"        "caci"        "caec"        "caecc"       "caipa"      
#> [111] "cames"       "camrsd"      "camsf"       "caricom"     "carifta"    
#> [116] "carii"       "catc"        "cbfp"        "cbi"         "cbss"       
#> [121] "ccnr"        "ccom"        "ccpa"        "ccomm"       "cdb"        
#> [126] "ceao"        "cec"         "ceepn"       "cefta"       "cei"        
#> [131] "celc"        "cemac"       "cento"       "cepgl"       "cern"       
#> [136] "cfatf"       "cfc"         "chstea"      "cica"        "cifc"       
#> [141] "cilss"       "cima"        "cis"         "cmaec"       "cmaoc"      
#> [146] "cmea"        "cmhasg"      "coe"         "colombo"     "comesa"     
#> [151] "confejes"    "coptac"      "cosave"      "cpab"        "cpsc"       
#> [156] "cpu"         "cslf"        "csto"        "ctcaf"       "cto"        
#> [161] "cwgc"        "cxc"         "cfrn"        "comab"       "comsec"     
#> [166] "corg"        "d8"          "dbgls"       "dlcoea"      "danube"     
#> [171] "eac"         "eacm"        "eacso"       "eadb"        "eaec"       
#> [176] "eapc"        "eapo"        "ebrd"        "ecb"         "ecca"       
#> [181] "eccas"       "eccb"        "eccd"        "eccm"        "eccpif"     
#> [186] "eco"         "ecowas"      "ecpta"       "ecsc"        "eec"        
#> [191] "efcc"        "efilwc"      "efta"        "eib"         "eip"        
#> [196] "eipa"        "eldo"        "emb"         "embc"        "embl"       
#> [201] "emi"         "emppo"       "epa"         "epfsc"       "epo"        
#> [206] "epu"         "eria"        "esa"         "eso"         "esro"       
#> [211] "etf"         "eu"          "eufmd"       "euramet"     "euratom"    
#> [216] "eurocontrol" "eurofima"    "euromet"     "entente"     "fao"        
#> [221] "fdiplac"     "fec"         "g15"         "g24"         "g3"         
#> [226] "gatt"        "gbact"       "gcc"         "gcrsnc"      "gef"        
#> [231] "geo"         "ghsi"        "glacsec"     "goic"        "hcpil"      
#> [236] "iabe"        "iabath"      "iacb"        "iaci"        "iacs"       
#> [241] "iacss"       "iacw"        "iadb"        "iadefb"      "iaea"       
#> [246] "iafc"        "iahc"        "iaias"       "iaic"        "iaigc"      
#> [251] "iaii"        "ialong"      "iamlo"       "iaphy"       "iara"       
#> [256] "iarhc"       "iaradio"     "iaruhr"      "ias"         "iasaj"      
#> [261] "iatb"        "iatsj"       "iattc"       "iba"         "ibcs"       
#> [266] "ibe"         "ibec"        "ibi"         "ibier"       "ibpmp"      
#> [271] "ibrd"        "icac"        "icai"        "icao"        "icamo"      
#> [276] "icc"         "iccec"       "iccilmb"     "icco"        "iccrom"     
#> [281] "iccs"        "iccslt"      "icdr"        "ices"        "icfam"      
#> [286] "ichrb"       "icmmp"       "icnc"        "icnwaf"      "icprp"      
#> [291] "icptu"       "icri"        "icrpbc"      "icse"        "icseaf"     
#> [296] "icsg"        "ictm"        "icfo"        "ichemo"      "icivdo"     
#> [301] "icomo"       "idc"         "iea"         "iec"         "ies"        
#> [306] "iexb"        "ifad"        "ifc"         "ifca"        "igad"       
#> [311] "igc"         "igcc"        "iho"         "iia"         "iicom"      
#> [316] "iie"         "iif"         "iifeo"       "iiwee"       "ijo"        
#> [321] "ilo"         "ilzsg"       "imbslav"     "imc"         "imf"        
#> [326] "imi"         "imo"         "imso"        "incap"       "infofish"   
#> [331] "infsmk"      "inpfc"       "inro"        "insg"        "intelsat"   
#> [336] "interpol"    "ioathre"     "iocom"       "ioez"        "iolm"       
#> [341] "iomig"       "iooc"        "iopcf"       "ioph"        "iorarc"     
#> [346] "iocc"        "ipc"         "ipi"         "ipedi"       "ipentc"     
#> [351] "iphyl"       "iprizec"     "irena"       "irlcs"       "iro"        
#> [356] "irsg"        "iru"         "isa"         "isb"         "isdb"       
#> [361] "ishrest"     "isrbc"       "isupt"       "isuc"        "itc"        
#> [366] "itcc"        "itcle"       "itpa"        "itro"        "itto"       
#> [371] "itu"         "itvrc"       "iuic"        "iupct"       "iupip"      
#> [376] "iuplaw"      "iupnvp"      "iupr"        "ivwo"        "iwsg"       
#> [381] "iwhale"      "iocean"      "jalaao"      "jinr"        "jnolcrh"    
#> [386] "lacac"       "lacp"        "laeo"        "lafdo"       "lafta"      
#> [391] "laia"        "laiec"       "latin"       "lcbc"        "lgida"      
#> [396] "loas"        "lon"         "maocn"       "marri"       "mcpttc"     
#> [401] "mcwcasm"     "mfo"         "miga"        "mru"         "mwn"        
#> [406] "mercosur"    "montreal"    "nacap"       "nafo"        "nafta"      
#> [411] "nam"         "nappo"       "nasco"       "nato"        "ncm"        
#> [416] "ncrr"        "nctr"        "ndf"         "neafc"       "nerc"       
#> [421] "nib"         "npafc"       "npfsc"       "npi"         "nrc"        
#> [426] "ntsc"        "nvc"         "nwhf"        "nordc"       "oapec"      
#> [431] "oas"         "oau"         "ocam"        "ocas"        "occedca"    
#> [436] "ocr"         "oecd"        "oecs"        "oeec"        "oic"        
#> [441] "oiv"         "omdkr"       "omvg"        "opanal"      "opec"       
#> [446] "osce"        "oslo"        "ospar"       "otif"        "pahc"       
#> [451] "paho"        "paigh"       "pap"         "papu"        "pc"         
#> [456] "pca"         "pcb"         "pcsp"        "ped"         "piarc"      
#> [461] "pibac"       "pices"       "pics"        "pif"         "pipd"       
#> [466] "pmaesa"      "psnarco"     "ptasea"      "puasp"       "rascom"     
#> [471] "rcaela"      "rcc"         "rcfc"        "recsa"       "rimmo"      
#> [476] "rioppah"     "riogroup"    "radiou"      "repcom"      "saafa"      
#> [481] "saarc"       "sacep"       "sacu"        "sadc"        "sadcc"      
#> [486] "sami"        "sartc"       "sca"         "scaf"        "sch"        
#> [491] "schengen"    "sco"         "seameo"      "seato"       "segib"      
#> [496] "sela"        "sica"        "sieca"       "sittdec"     "spc"        
#> [501] "srdo"        "swapu"       "swpd"        "sugu"        "tcrmg"      
#> [506] "turksoy"     "uasc"        "ubec"        "udeac"       "uemoa"      
#> [511] "uiucv"       "ukdwd"       "um"          "umac"        "umoa"       
#> [516] "un"          "unesco"      "unido"       "unidroit"    "upu"        
#> [521] "usp"         "valdivia"    "vasab"       "waec"        "wahc"       
#> [526] "waho"        "wcdc"        "wco"         "weu"         "who"        
#> [531] "wipo"        "wmo"         "wnf"         "wpact"       "wto"        
#> [536] "wtouro"      "wassen"     
state_year_format3 %>%
  select(ccode:afgec) %>%
  filter(year > 1990) %>%
  glimpse()
#> Rows: 4,567
#> Columns: 26
#> $ ccode    <dbl> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2…
#> $ year     <dbl> 1991, 1992, 1993, 1994, 1995, 1996, 1997, 1998, 1999, 2000, 2…
#> $ state    <chr> "usa", "usa", "usa", "usa", "usa", "usa", "usa", "usa", "usa"…
#> $ aaaid    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aacb     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aalco    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aaro     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aata     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aatpo    <dbl> 0, 0, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,…
#> $ abeda    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abepseac <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acc      <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acct     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acdt     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ aci      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -1, -1, -1, -1, -1, -1, -1, -1…
#> $ acml     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -1, -1, -1, -1, -1, -1, -1, -1,…
#> $ acp      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acpeu    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acs      <dbl> -1, -1, -1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acso     <dbl> -9, -9, -9, -9, -9, -9, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acssrb   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acu      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acwl     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 0, 0, 0, 0, 0, 0, 0, …
#> $ afesd    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ afeximb  <dbl> -1, -1, -9, -9, -9, -9, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, -…
#> $ afgec    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…

# Recode numerical to factors: with a sample
sample_state_year <- state_year_format3 %>%
  as_tibble() %>%
  select(ccode:afgec) %>%
  filter(year == 1990)

sample_state_year %>% glimpse()
#> Rows: 161
#> Columns: 26
#> $ ccode    <dbl> 2, 20, 31, 40, 41, 42, 51, 52, 53, 54, 55, 56, 57, 58, 60, 70…
#> $ year     <dbl> 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1…
#> $ state    <chr> "usa", "canada", "bahamas", "cuba", "haiti", "domrepublic", "…
#> $ aaaid    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aacb     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aalco    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aaro     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aata     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ aatpo    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abeda    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ abepseac <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acc      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acct     <dbl> 0, 1, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acdt     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ aci      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acml     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acp      <dbl> 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 1, 0, 0, 0, 0…
#> $ acpeu    <dbl> 0, 0, 1, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0…
#> $ acs      <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acso     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ acssrb   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acu      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ acwl     <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ afesd    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ afeximb  <dbl> -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -…
#> $ afgec    <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…

# Recode
sample_state_year_recoded <- sample_state_year %>%
  mutate(across(-c(ccode:state), igo_recode_stateyear))

sample_state_year_recoded %>% glimpse()
#> Rows: 161
#> Columns: 26
#> $ ccode    <dbl> 2, 20, 31, 40, 41, 42, 51, 52, 53, 54, 55, 56, 57, 58, 60, 70…
#> $ year     <dbl> 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1990, 1…
#> $ state    <chr> "usa", "canada", "bahamas", "cuba", "haiti", "domrepublic", "…
#> $ aaaid    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aacb     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aalco    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aaro     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aata     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ aatpo    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ abeda    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ abepseac <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acc      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acct     <fct> No Membership, Full Membership, No Membership, No Membership,…
#> $ acdt     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ aci      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acml     <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acp      <fct> No Membership, No Membership, Full Membership, No Membership,…
#> $ acpeu    <fct> No Membership, No Membership, Full Membership, No Membership,…
#> $ acs      <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acso     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ acssrb   <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acu      <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ acwl     <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ afesd    <fct> No Membership, No Membership, No Membership, No Membership, N…
#> $ afeximb  <fct> IGO Not In Existence, IGO Not In Existence, IGO Not In Existe…
#> $ afgec    <fct> No Membership, No Membership, No Membership, No Membership, N…