State System Membership (v2016)

The list of states with COW abbreviations and ID numbers, plus the field state from state_year_format3.

Format

data frame with 243 rows:

ccode: COW country number.

stateabb: COW state abbreviation (3 characters).

statenme: COW state name.

styear...endday: Fields to identify the begining and the end of each tenure.

version: Data file version number.

state: Abbreviated state name as appear in state_year_format3.

Source

State System Membership (v2016), The Correlates of War Project.

Details

This data set contains the list of states in the international system as updated and distributed by the Correlates of War Project.

This data set contains the list of states in the international system as updated and distributed by the Correlates of War Project. These data sets identify states, their standard Correlates of War "country code" or state number (used throughout the Correlates of War project data sets), state abbreviations, and dates of membership as states and major powers in the international system.

The Correlates of War project includes a state in the international system from 1816-2016 for the following criteria:

Prior to 1920 the entity must have had a population greater than 500,000 and have had diplomatic missions at or above the rank of charge d'affaires with Britain and France.

After 1920 the entity must be a member of the League of Nations or the United Nations, or have a population greater than 500,000 and receive diplomatic missions from two major powers.

Note

state variable added to original data to help comparison across datasets on this package.

References

Correlates of War Project. 2017 "State System Membership List, v2016." Online, https://correlatesofwar.org/.

See also