State System Membership (v2016)

The list of states with COW abbreviations and ID numbers, plus the field state from state_year_format3.

Format

data frame with 243 rows. Relevant fields:

ccode: COW country number.

stateabb: COW state abbreviation (3 characters).

statenme: COW state name.

styear...endday: Fields to identify the begining and the end of each tenure.

version: Data file version number.

state: Abbreviated state name as appear in state_year_format3.

Source

State System Membership (v2016), The Correlates of War Project.

Details

This data set contains the list of states in the international system as updated and distributed by the Correlates of War Project.

This data set contains the list of states in the international system as updated and distributed by the Correlates of War Project. These data sets identify states, their standard Correlates of War "country code" or state number (used throughout the Correlates of War project data sets), state abbreviations, and dates of membership as states and major powers in the international system.

The Correlates of War project includes a state in the international system from 1816-2016 for the following criteria:

Prior to 1920 the entity must have had a population greater than 500,000 and have had diplomatic missions at or above the rank of charge d'affaires with Britain and France.

After 1920 the entity must be a member of the League of Nations or the United Nations, or have a population greater than 500,000 and receive diplomatic missions from two major powers.

Note

state variable added to original data to help comparison across datasets on this package.

References

Correlates of War Project. 2017 "State System Membership List, v2016." Online, https://correlatesofwar.org/.

See also

Examples

# example code
data("states2016")
dplyr::tibble(states2016)
#> # A tibble: 243 × 11
#>    ccode stateabb statenme  styear stmonth stday endyear endmonth endday version
#>    <int> <chr>    <chr>      <int>   <int> <int>   <int>    <int>  <int>   <int>
#>  1     2 USA      United S…   1816       1     1    2016       12     31    2016
#>  2    20 CAN      Canada      1920       1    10    2016       12     31    2016
#>  3    31 BHM      Bahamas     1973       7    10    2016       12     31    2016
#>  4    40 CUB      Cuba        1902       5    20    1906        9     25    2016
#>  5    40 CUB      Cuba        1909       1    23    2016       12     31    2016
#>  6    41 HAI      Haiti       1859       1     1    1915        7     28    2016
#>  7    41 HAI      Haiti       1934       8    15    2016       12     31    2016
#>  8    42 DOM      Dominica…   1894       1     1    1916       11     29    2016
#>  9    42 DOM      Dominica…   1924       9    29    2016       12     31    2016
#> 10    51 JAM      Jamaica     1962       8     6    2016       12     31    2016
#> # ℹ 233 more rows
#> # ℹ 1 more variable: state <chr>